Fix golang lint version

+++ internal/worker/worker.go
@@ -168,7 +168,7 @@ func (w *Worker) retryJob(ctx context.Context, jobID uuid.UUID, errMsg string, a
 	if exp > 10 {
 		exp = 10
 	}
-	backoff := time.Duration(1<<uint(exp)) * time.Second
+	backoff := time.Duration(1<<exp) * time.Second // exp is capped at 10 above
 	_, _ = w.db.Exec(ctx, `
 		UPDATE jobs SET status = 'pending', last_error = $1, run_at = now() + $2::interval, locked_by_node_id = NULL, locked_at = NULL, updated_at = now()
 		WHERE id = $3
